Add the FVT test for the ovis_ev library
The core logic that performs the verification is implemented in the C
program (C/test_ovis_ev.c). The Python test script (test_ovis_ev) starts
a TADA Docker Cluster and executes the C test program and then reads as
well as sends the test results to the TADA server.