Introduce the test_store plugin
The patch introduces the test_store store plugin, which is an LDMSD
store plugin. Test scripts could use the plugin to store data
instead of any store plugins in the OVIS source tree so that the tests
will be independent of any changes in the store plugins in the source
tree.