Remove `--latest` option from tadaq and opt for `--all`
Before this patch, the tadaq default behavior was to list all results,
including the previous runs of the same test ID. The user had to specify
`--latest` to list only the latest results. This was cumbersome.

The new behavior is to list only the latest results (of each test ID),
and the user has to specify `--all` to include the previous results.