tadaq --only-STATUS shall skip tests with no macthing assertions
`tadaq --only-STATUS` (e.g. `tadaq --only-failed`) only prints
assertions that match the STATUS. If the test does not have any
assertions matching the status (e.g. `--only-failed` but we only have
passed assertions), `tadaq` printed the information of the test
with an empty assertion list, which can be confusing. This patch skip
the print of the test if the matching assertion list is empty.