tadaq query with commit-id from the latest test by default
tadaq queried all commit IDs by default, which produced quite a large
output. This patch restricted the query commit-id to the commit-id of
the most-recent test by the caller (as if
`--commit-id=<COMMIT-ID-OF-THE-LATEST-RUN>` is given). The users can
provide `--commit-id=*` to get test results from all commit IDs (match
any commit IDs).