Fix `LDMS_Test.LDMSDCluster.sbatch()` environment
We used to be able to directly invoke the program (sbatch in this case)
and the basic environment variables (e.g. PATH) would be conveniently
setup as if we called the program under `/bin/bash`. With the recent
docker update, it seemed that the basic environment variables were not
setup properly. Hence, this patch explicitly put the `sbatch` call under
`/bin/bash`.