Improve `spank_notifier_test` stability
`spank_notifier_test` restarts (stop-and-start) `slurmd` and `slurmctld`
in the middle of the test. The system resources (e.g. ports) could still
be not completely cleaned up when the new process start, causing the
test to fail. This patch makes the test more stable by checking the
start/stop conditions and adding some time between the stop-start.