Set pty size in singularity runtime to avoid autowrap
By default, the pty size of the `Container.exec_interact()` in
singularity runtime is automatically set to be the same size as the
current shell pty. As a result, a long command will automatically
wrapped into a new line with a bunch of terminal io control-sequence
characters. This hinders interactive output result comparison. This
patch set the PTY size to be 80 rows x 240 columns to avoid the line
wrapping.