ldmsd_stream_test small wait after subscriber start
On cygnus, after the subscriber daemon start command had issued, the
publishers commans were also issued right after it. This caused
occasional missing data error because after the
subscriber-daemon-start-command had returned, it did not mean that the
daemon itself was ready to take data. This patch add a little wait
time between the subscriber and the publisher command to allow the
subscriber daemon to be up and ready.