prdcr_subscribe_test small wait before reading output files
After the data was published, we issued the command to read the data
right away. On cygnus cluster, the data had not arrived to the output
file in time and caused data verification error. This patch adds a
small wait time after the data publishing to allow the data to be
flushed into the output files.