LDMSDCluster.ssh_keyscan() docker stderr fix
docker `exec_run` globs stdout and stderr together and ignores stderr
redirection request. This creates a problem with `ssh-keyscan` command
as stderr stream corrupts stdout data (e.g. put '# comment' in the
middle of the stdout line which describing a remote key). This patch
works around the issue by putting the `ssh-keyscan` command under
`bash` so that the redirection directive is honored.