Increase wait time in maestro_raft_test after maestro resurrection
The wait time after maestro resurrection was 2 seconds. This was a
borderline since the test sometimes failed because the resurrected
maestro was still in the `CANDIDATE` state (it was still trying to join
the group). This patch increases the check wait time after
the resurrection to 10 seconds to increase the likelihood that the
resurrected maestro successfully join the group (if the resurrected
maestro could not successfully join the group within 10 second in this
test environment, we should investigate the cause). The result has been
consistent on Cygnus.