Update `ldms_stream_test`
`ldms_stream_publish()` now uses stream name as a part to determine
rail endpoint index from the get go (instead of always using ep[0] on
the first hop). This broke `ldms_stream_test` in the stream statistic
tracking part due to different port number of the endpoint being sent on
the first hop. This patch made the first hop to have a 1-endpoint rail
so that the existing test can still track the stream statistics easily.