Add `atstart` script support in ldms-test config file
Add `atstart` support in `[ldms-test]` section in the config file. If
`atstart` is defined, the script will be copied to `/atstart.sh` in the
container and executed after the container started.

Example:
```
[ldms-test]
atstart = myscript.sh
```

In this example, `myscript.sh` is copied to `/atstart.sh` in the
container and will be executed when the container started.